Refactored tests into unit and integration.
