+++ mapstory/templates/site_base.html
@@ -1,3 +1,4 @@
+<!DOCTYPE html>
 <html lang="en">
     <head>
         <meta charset="utf-8">
@@ -16,7 +17,7 @@
         {% if site.assets.favicon %}
         <link rel="shortcut icon" href="{{ MEDIA_URL }}{{ site.assets.favicon.name }}">
         {% endif %}
-        {% if GOOGLE_ANALYTICS %}{% include "mapstory/_google_analytics.html" %}{% endif %}  
+        {% if GOOGLE_ANALYTICS %}{% include "mapstory/_google_analytics.html" %}{% endif %}
         {% block extra_head %}{% endblock %}
     {% endblock head %}
     </head>
@@ -28,7 +29,7 @@
         <div class="alert alert-danger alert-dismissible" role="alert" id="ieflag" style="display:none">
             <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">&times;</span></button>
             <h1>You are using an outdated browser that is not supported by GeoNode.</h1>
-            <p>Please use a <strong>modern browser</strong> like Mozilla Firefox, Google Chrome or Safari.</p>  
+            <p>Please use a <strong>modern browser</strong> like Mozilla Firefox, Google Chrome or Safari.</p>
         </div>
         <div id="wrap">
             {% block middle %}
