+++ mapstory/tests/js/icon_upload.po.js
@@ -7,10 +7,10 @@
 require('./waitReady.js');
 var path = require('path');
 
-const testPNGRelative = '../test_assets/icon.png';
+const testPNGRelative = '../sampledata/icon.png';
 const testPNGFile = path.resolve(__dirname, testPNGRelative);
 
-const testSVGRelative = '../test_assets/icon-github.svg';
+const testSVGRelative = '../sampledata/icon-github.svg';
 const testSVGFile = path.resolve(__dirname, testSVGRelative);
 
 var iconUploadWizard = function() {
