+++ mapstory/tests/sampledata/icon.png
