+++ mapstory/tests/unit/test_HelperMethods.py
@@ -1,0 +1,82 @@
+from unittest import skip
+from django.test import TestCase
+
+from mapstory.models import _stamp
+from mapstory.models import GroupProfile
+from mapstory.models import name_post_save
+from mapstory.models import get_images
+from mapstory.models import get_sponsors
+from mapstory.models import get_featured_groups
+from mapstory.models import get_group_layers
+from mapstory.models import get_group_maps
+
+class MockInstance(object):
+    """
+    Helper Mock Instance
+    """
+    name = 'A name'
+
+class TestHelperMethods(TestCase):
+    """
+    Model helper methods tests
+    """
+    def test_stamp(self):
+        """
+        Should return a hash
+        """
+        self.assertIsNotNone(_stamp("Hola!"))
+        self.assertEqual(_stamp("123"),_stamp("123"))
+        self.assertNotEqual(_stamp("123"), _stamp("321"))
+
+    def test_stamp_params(self):
+        """
+        Should only stamp strings or buffers
+        """
+        self.assertRaises(TypeError, _stamp, 123)
+
+    def test_stamp_returns(self):
+        """
+        Should be 8 in length
+        """
+        self.assertEqual(len(_stamp("456")), 8)
+        self.assertEqual(len(_stamp("45685739391")), 8)
+        self.assertEqual(len(_stamp("")), 8)
+        self.assertEqual(len(_stamp("AaaaaaaaaAAaaAAAAaaA")), 8)
+
+    def test_name_post_save(self):
+        """
+        name_post_save() tests
+        """
+        name_post_save(MockInstance())
+
+    def test_get_images(self):
+        self.assertIsNotNone(get_images())
+        # Should return an array
+        self.assertIsNotNone(get_images().count())
+        self.assertEqual(get_images().count(), 0)
+
+    def test_get_sponsors(self):
+        sponsors = get_sponsors()
+        self.assertIsNotNone(sponsors)
+        self.assertEqual(sponsors.count(), 0)
+
+    def test_get_featured_groups(self):
+        self.assertEqual(get_featured_groups().count(), 0)
+
+    @skip("TODO")
+    def test_group_layers(self):
+        gp = GroupProfile()
+        self.assertEqual(get_group_layers(gp).count(), 0)
+    
+    @skip("TODO")
+    def test_group_maps(self):
+        gp = GroupProfile()
+        self.assertEqual(get_group_maps(gp).count(), 0)
+
+    @skip("TODO")
+    def test_group_journals(self):
+        pass
+
+    @skip("TODO")
+    def test_mapstory_profile_post_save(self):
+        pass
