+++ mapstory/tests/unit/test_Initiatives.py
@@ -4,8 +4,8 @@ from datetime import datetime
 from geonode.groups.models import GroupProfile
 from geonode.contrib.collections.models import Collection
 
-from .MapStoryTestMixin import MapStoryTestMixin
-from .AdminClient import AdminClient
+from mapstory.tests.MapStoryTestMixin import MapStoryTestMixin
+from mapstory.tests.AdminClient import AdminClient
 
 
 class InitiativeTests(MapStoryTestMixin):
