+++ mapstory/tests/unit/test_Organizations.py
@@ -1,36 +1,52 @@
-from django.contrib.auth import get_user_model, authenticate
-from django.test import TestCase, Client
-from django.test.utils import override_settings
-from geonode.geoserver.helpers import gs_catalog
-from socket import error as socket_error
-from geonode.base.models import TopicCategory
-from geonode.base.populate_test_data import create_models
-from .AdminClient import AdminClient
-from django.core.urlresolvers import reverse
-from geonode.groups.models import GroupProfile, GroupMember
-from geonode.contrib.collections.models import Collection
 from datetime import datetime
 from unittest import skip
-from bs4 import BeautifulSoup
-from mapstory.tests.MapStoryTestMixin import MapStoryTestMixin
 
+from bs4 import BeautifulSoup
+
+from django.core.urlresolvers import reverse
+from django.test import TestCase, Client
+from django.test.utils import override_settings
+
+from geonode.contrib.collections.models import Collection
+from geonode.groups.models import GroupProfile
+
+from mapstory.views import organization_create, organization_edit, organization_detail, organization_members
+from mapstory.tests.MapStoryTestMixin import MapStoryTestMixin
+from mapstory.tests.AdminClient import AdminClient
+
+class TestUnitOrganizations(TestCase):
+    def test_view_imports(self):
+        self.assertIsNotNone(organization_create, "Failed to import 'organization_create'")
+        self.assertIsNotNone(organization_edit, "Failed to import 'organization_edit'")
+        self.assertIsNotNone(organization_detail, "Failed to import 'organization_detail'")
+        self.assertIsNotNone(organization_members, "Failed to import 'organization_members'")
+
+    def test_url_api(self):
+        self.assertEqual(reverse('organization_create'), u'/organizations/create/')
+        self.assertEqual(reverse('organization_edit', kwargs={'slug': 'testslug'}), u'/organizations/edit/testslug')
+        self.assertEqual(reverse('organization_detail', kwargs={'slug': 'testslug'}), u'/organizations/testslug')
+        self.assertEqual(reverse('organization_members', kwargs={'slug': 'testslug'}), u'/organizations/members/testslug')
+
+    def test_create_unauthorized(self):
+        # TODO: Test creating with guest user
+        pass
+
+    def test_create_authorized(self):
+        # TODO: Test creating 
+        # c = Client()
+        pass
+        # c.put(reverse('organization_create'), )
+
+    def test_create_template(self):
+        # TODO: Figure out the template
+        # Should use the correct template
+        c = Client()
+        response = c.get(reverse('index_view'))
+        self.assertContains(response, "<!DOCTYPE html>", count=1, status_code=200, html=False)
+        # self.assertTemplateUsed()
+        # 
 
 class MapStoryOrganizationTests(MapStoryTestMixin):
-    """Organization Tests
-
-    Organizations API
-    -----------------
-        ```
-        url(r'^organizations/create/$', organization_create, name='organization_create'),
-        url(r'^organizations/(?P<slug>[^/]*)$', organization_detail, name='organization_detail'),
-        url(r'^organizations/edit/(?P<slug>[^/]*)$', organization_edit, name='organization_edit'),
-        url(r'^organizations/members/(?P<slug>[^/]*)$', organization_members, name='organization_members'),
-        url(r'^organizations/invite/(?P<slug>[^/]*)$', organization_invite, name='organization_invite'),
-        url(r'^organizations/(?P<slug>[^/]*)/members_add/$', organization_members_add, name='organization_members_add'),
-        url(r'^organizations/(?P<slug>[^/]*)/member_remove/(?P<username>.+)$', organization_member_remove, name='organization_member_remove'),
-        ```
-    """
-
     def setUp(self):
         self.username, self.password = self.create_user('admin', 'admin', is_superuser=True)
         self.non_admin_username, self.non_admin_password = self.create_user('non_admin', 'non_admin')
@@ -339,4 +355,3 @@ class MapStoryOrganizationTests(MapStoryTestMixin):
 
         # Should have 27 fields total
         self.assertEqual(len(soup.find_all('input')), 28)
-
