+++ mapstory/tests/unit/test_unit_model_sponsor.py
@@ -1,0 +1,8 @@
+from django.test import TestCase
+
+from mapstory.models import Sponsor
+
+class SponsorModelTest(TestCase):
+    def test_sponsor_name(self):
+        sponsor = Sponsor(name='My Sponsor Name')
+        self.assertEqual(unicode(sponsor), 'Sponsor - ' + sponsor.name)
